move alternate package autocomplete to jsonapicontroller
