+++ src/NuGetGallery/RouteName.cs
@@ -45,7 +45,6 @@ namespace NuGetGallery
         public const string PackageAction = "PackageAction";
         public const string PackageDeleteAction = "PackageDeleteAction";
         public const string PackageDeprecateAction = "PackageDeprecateAction";
-        public const string PackageDeprecationGetAlternatePackageVersionsAction = "PackageDeprecationGetAlternatePackageVersions";
         public const string PushPackageApi = "PushPackageApi";
         public const string PublishPackageApi = "PublishPackageApi";
         public const string DeletePackageApi = "DeletePackageApi";
